requirements.txt
click==7.0
gradient>=0.6
matplotlib>=3.1
multipledispatch==0.6.0
numpy==1.17.3
pandas>=0.25
Pillow==6.2.1
plotly>=4.2
pytest>=5.2
pytest-arraydiff==0.3
python-dotenv==0.10.3
pytorch-ignite>=0.2
pyyaml==5.1.2
scikit-learn==0.21.3
tb-nightly
torch>=1.3
torch-cluster>=1.4
torch-geometric>=1.3
torch-scatter>=1.3
torch-sparse>=0.4
torch-spline-conv>=1.1
torchvision>=0.4
tqdm>=4.36
